Fix legend yerr (#40777)
